updated transfer page
